refactor capture
