--- src/clients/paymentCapture/capture/types.ts
@@ -1,16 +1,0 @@
-import type { MercadoPagoConfig } from '@src/mercadoPagoConfig';
-
-export declare type Capture = {
-  transaction_amount?: number;
-};
-
-export declare type PaymentCaptureRequest = {
-  id: string;
-  body: Capture;
-}
-
-export declare type PaymentCapture = {
-  id: string;
-  body: Capture;
-  config: MercadoPagoConfig;
-};
