+++ src/clients/payments/capture/types.ts
@@ -1,0 +1,10 @@
+import type { MercadoPagoConfig } from '@src/mercadoPagoConfig';
+
+export declare type PaymentCaptureRequestBody = {
+  id: string;
+  transaction_amount: number,
+}
+
+export declare interface PaymentCaptureRequest extends PaymentCaptureRequestBody {
+  config: MercadoPagoConfig;
+}
